Add files via upload
The hotel booking data analysis project is based on real-world data recorded from 2015 to 2017 for a city and a resort hotel. The project aims to explore the dataset from both the hotel and customer perspectives, with a particular focus on understanding and visualizing reasons for booking cancellations across various parameters. The dataset includes a range of details such as guest information, bookings, and cancellations, which will be analyzed to gain insights into patterns and trends in hotel bookings and cancellations.